refactor: hotel-booking
